requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
PyMySQL==1.1.0
python-dotenv
requests==2.25.1
beautifulsoup4==4.12.2
Pillow==9.0.1
boto3==1.28.64